+++ ChangeLog.md
@@ -77,6 +77,8 @@
 
   * Add docs for external grading statistics (Matt West).
 
+  * Add ability to restore original file in `pl_file_editor` (Nathan Walters).
+
   * Fix broken file upload element (Nathan Walters).
 
   * Fix broken popover and improve assessment label styles (Nathan Walters).